refacotr(LottoCompany, LottoBundle): LottoCompany 과도한 역할 분리
- LottoCompany 에서 생성하는 LottoBundle 역할 LottoBundle로 옮기기
