+++ src/main/java/lotto/domain/lotto/LottoBundle.java
@@ -1,6 +1,11 @@
 package lotto.domain.lotto;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class LottoBundle {
 
@@ -8,6 +13,34 @@ public class LottoBundle {
 
     public LottoBundle(List<Lotto> lottoList) {
         this.lottoList = lottoList;
+    }
+
+    public LottoBundle(int count) {
+        List<Lotto> lottoList = new ArrayList<>();
+        for (int i = 0; i < count; i++) {
+            lottoList.add(makeRandomLotto());
+        }
+//        return new LottoBundle(lottoList);
+        this.lottoList = lottoList;
+    }
+
+    private Lotto makeRandomLotto() {
+        List<Integer> lottoNumberBox = makeLottoBox();
+        Collections.shuffle(lottoNumberBox);
+
+        return new Lotto(makeLottoNumber(lottoNumberBox.subList(0, 6)));
+    }
+
+    private List<Integer> makeLottoBox() {
+        return IntStream.range(1, 45)
+                .boxed()
+                .collect(Collectors.toList());
+    }
+
+    private Set<LottoNumber> makeLottoNumber(List<Integer> numberList) {
+        return numberList.stream()
+                .map(number -> LottoNumber.of(number))
+                .collect(Collectors.toSet());
     }
 
     public List<Lotto> unfoldLottoBundle() {
