+++ src/main/java/lotto/domain/lotto/LottoCompany.java
@@ -13,7 +13,7 @@ import java.util.stream.Stream;
 public class LottoCompany {
 
     private static final String SPLIT_DELIMITER = ",";
-    private static LottoCompany lottoCompany = new LottoCompany();
+    private static final LottoCompany lottoCompany = new LottoCompany();
 
     private LottoCompany() {};
 
@@ -32,39 +32,12 @@ public class LottoCompany {
         return new ManualLottoBundle(lottoList);
     }
 
-    public LottoBundle makeLottoBundle(int count) {
-        List<Lotto> lottoList = new ArrayList<>();
-        for (int i = 0; i < count; i++) {
-            lottoList.add(makeRandomLotto());
-        }
-        return new LottoBundle(lottoList);
-    }
-
     private Lotto convertToLotto(String stringNumbers) {
         if (stringNumbers == null) {
             throw new IllegalArgumentException("로또로 전환할 번호들을 입력하지 않으셨습니다.");
         }
         String[] splitedNumbers = splitNumbers(stringNumbers);
         return new Lotto(makeLottoNumber(splitedNumbers));
-    }
-
-    private List<Integer> makeLottoBox() {
-        return IntStream.range(1, 45)
-                .boxed()
-                .collect(Collectors.toList());
-    }
-
-    private Lotto makeRandomLotto() {
-        List<Integer> lottoNumberBox = makeLottoBox();
-        Collections.shuffle(lottoNumberBox);
-
-        return new Lotto(makeLottoNumber(lottoNumberBox.subList(0, 6)));
-    }
-
-    private Set<LottoNumber> makeLottoNumber(List<Integer> numberList) {
-        return numberList.stream()
-                .map(number -> LottoNumber.of(number))
-                .collect(Collectors.toSet());
     }
 
     private String[] splitNumbers(String winnerNumbers) {
